requirements.txt
Django==3.0.4
psycopg2==2.8.4
django-rq-scheduler==1.1.3
django-background-tasks==1.2.5